# openapispec.yaml
openapi: 3.0.0
info:
  title: SOA I projekat
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /get:
    get:
      tags:
        - Maternal Health Risks
      summary: get
      parameters:
        - name: perPage
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /create:
    post:
      tags:
        - Maternal Health Risks
      summary: create
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Age: 9
                SystolicBP: 9
                DiastolicBP: 9
                BS: 9
                BodyTemp: 9
                HeartRate: 9
                RiskLevel: high risk
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /update/{id}:
    put:
      tags:
        - Maternal Health Risks
      summary: update
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                Age: 9
                SystolicBP: 6
                DiastolicBP: 9
                BS: 9
                BodyTemp: 9
                HeartRate: 9
                RiskLevel: high risk
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /delete/{id}:
    delete:
      tags:
        - Maternal Health Risks
      summary: delete
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}